Add sled provision data to sled page
