+++ app/pages/system/inventory/sled/SledPage.tsx
@@ -19,6 +19,8 @@ import { PropertiesTable } from '~/ui/lib/PropertiesTable'
 import { truncate } from '~/ui/lib/Truncate'
 import { pb } from '~/util/path-builder'
 
+import { ProvisionPolicyBadge, SledKindBadge, SledStateBadge } from './SledBadges'
+
 export async function clientLoader({ params }: LoaderFunctionArgs) {
   const { sledId } = requireSledParams(params)
   await apiQueryClient.prefetchQuery('sledView', {
@@ -47,23 +49,29 @@ export default function SledPage() {
         <PropertiesTable.Row label="sled id">
           <span className="text-default">{sled.id}</span>
         </PropertiesTable.Row>
+        <PropertiesTable.Row label="kind">
+          <SledKindBadge policy={sled.policy} />
+        </PropertiesTable.Row>
         <PropertiesTable.Row label="part">
           <span className="text-default">{sled.baseboard.part}</span>
+        </PropertiesTable.Row>
+        <PropertiesTable.Row label="provision policy">
+          <ProvisionPolicyBadge policy={sled.policy} />
         </PropertiesTable.Row>
         <PropertiesTable.Row label="serial">
           <span className="text-default">{sled.baseboard.serial}</span>
         </PropertiesTable.Row>
+        <PropertiesTable.Row label="state">
+          <SledStateBadge state={sled.state} />
+        </PropertiesTable.Row>
         <PropertiesTable.Row label="revision">
           <span className="text-default">{sled.baseboard.revision}</span>
         </PropertiesTable.Row>
-        <PropertiesTable.Row label="rack id">
-          <span className="text-default">{sled.rackId}</span>
-        </PropertiesTable.Row>
-        <PropertiesTable.Row label="location">
-          <span className="text-disabled">Coming soon</span>
-        </PropertiesTable.Row>
         <PropertiesTable.Row label="usable hardware threads">
           <span className="text-default">{sled.usableHardwareThreads}</span>
+        </PropertiesTable.Row>
+        <PropertiesTable.Row label="rack id">
+          <span className="text-default">{sled.rackId}</span>
         </PropertiesTable.Row>
         <PropertiesTable.Row label="usable physical ram">
           <span className="pr-0.5 text-default">{ram.value}</span>
